Add files via upload
This Python script, named 'pdf_reader.py', is designed to read a PDF version of a syllabus stored on your local machine. Currently, the file path is hardcoded, but it will be integrated with Flask in the future for dynamic file selection. The script extracts the contents of the PDF file and stores them in a database, such as SQLite or another preferred database system. I provided a copy of the syllabus that I used from my AI class. 

We need to work on the database section of this script once we decide on which database we are using for our project.